万象归一 Emerald（E）  ⚠ 重要特性  - Powered by Aethex Matrix ®：原生奥境智构矩阵重构应用 » 新增  - ⌈生活书⌋：生活记录工具，记录生活中的点滴故事、成长历程和美好时光。以时间为轴，串联起生活点滴，让回忆更加温馨美好 » 优化  - 主题渲染效果  - 多处细节优化

+++ app/src/main/java/com/wqz/allinone/act/note/NoteListActivity.kt
@@ -73,7 +73,7 @@ class NoteListActivity : ComponentActivity() {
 
         setContent {
             AllInOneTheme {
-                AppBackground.CirclesBackground {
+                AppBackground.BreathingBackground {
                     NoteListScreen()
                 }
             }
@@ -239,7 +239,7 @@ class NoteListActivity : ComponentActivity() {
                     horizontalAlignment = Alignment.CenterHorizontally
                 ) {
                     Text(
-                        text = stringResource(R.string.text_delete_confirm_hint),
+                        text = stringResource(R.string.delete_confirm_hint),
                         fontSize = 20.sp,
                         fontWeight = FontWeight.Bold,
                         maxLines = 1
@@ -251,10 +251,23 @@ class NoteListActivity : ComponentActivity() {
                         ItemX.Button(
                             text = stringResource(R.string.delete)
                         ) {
-                            showDialog = 2
-                            Toast.makeText(this@NoteListActivity, "删除成功", Toast.LENGTH_SHORT)
-                                .show()
-                            viewModel.deleteNoteWithDelay(note.id)
+                            if (!note.isLocked) {
+                                showDialog = 2
+                                Toast.makeText(
+                                    this@NoteListActivity,
+                                    "删除成功",
+                                    Toast.LENGTH_SHORT
+                                )
+                                    .show()
+                                viewModel.deleteNoteWithDelay(note.id)
+                            } else {
+                                Toast.makeText(
+                                    this@NoteListActivity,
+                                    "该笔记已锁定",
+                                    Toast.LENGTH_SHORT
+                                )
+                                    .show()
+                            }
                         }
                         Spacer(modifier = Modifier.width(10.dp))
                         ItemX.Button(
